config: Fix bug with reading cache line size option.
The easiest way to modify the system cache line size is to read it from
Options, rather than try to override it later if the value in gem5.cfg
is different. This means that it must be specified in the command that
runs gem5.

In addition, the DerivedO3CPU model asserts that its fetch buffer size
cannot be larger that the cache line size, so an additional check was
added for this.

Change-Id: I5f9f56c434dad084db264369f207a41585b38c69