aladdin: Enable memory traffic tracing.
This requires the probotuf library to be installed, and this brings the
Aladdin submodule pointer up to date with the necessary changes in
Aladdin as well.

To trace scratchpad and/or cache traffic on a per-accelerator basis,
include the attribute `record_memory_traffic = True` in the gem5.cfg
file for each accelerator, or mark it as default to trace all
accelerators in the system.

To trace all memory traffic going to/from main memory (which applies to
the entire system), add `--record-dram-traffic` to the command line
options (after aladdin_se.py).

Change-Id: I2a7ce01ef37155ca611972afe543239fb95c88d8