aladdin: Add defaults for config options.
In the past, every time a new config option was introduced, all
existing config files had to be updated with this new option or
gem5 would crash because the SafeConfigParser module does not provide
for defaults on a per get_*() call granularity. This commit updates the
aladdin_template.cfg file with default values for all optional options
as well as some documentation.

Change-Id: I06d9da7555a24d81d7d5e778c549ac24823f9c90