config: Update Aladdin datapath names.
The original code attached a list of datapaths to the System object, so
each individual datapath would be named "system.datapaths0",
"system.datapaths1", etc. This is both ugly and misleading, since in
other places, the datapath used to be called "datapathN" where N was the
accelerator ID. This commit changes the SimObject structure and updates
the naming to be "ACCELNAME_datapath".

Change-Id: Iee6ac5722effca46832c581ef066fc5ec4491556