stats: New implementation of SQL output.
The existing implementation was written in Python and relied on certain
SWIG cross-language polymorphism features, which PyBind11 lacks. The new
implentation is entirely written in C++, feature-complete with respect
to the original implementation, and compatible with PyBind11.

This requires the SQLite3 development headers and libraries to be
installed. SCons will try to find them automatically and if they are
found, it will enable SQL outputs. No build flags or options need to be
manually specified.

Change-Id: I308247a4d6923d0162ef12405210004f8cdf3fc6