arch-x86: Implement VEX.128 iforms of VCVTPS2PH/VCVTPH2PS.
These instructions enable conversions between single-precision and
half-precision floating point formats and comprise the crux of the F16C
ISA extension. We only implement the 64-bit and 128-bit iforms; as
gem5's x86 model does not support YMM registers, we cannot implement the
256-bit iforms.

Source and destination can be either registers or a 64-bit memory
location.  The conversion is done by an external library, FP16, included
in this commit. It does not support all the rounding modes, so the
immediate operand for rounding control is ignored.

Change-Id: I46a1404179bf278be1b70193bac04556c6688fe8