aladdin: Add ACP implementation in Ruby.
The current ACP implementation in the classic memory model requires a small
ACP cache, which has two downsides:
1) The ACP cache will buffer the ACP writes, as opposed to directly writing
data into the L2 cache. This can lead to extra latency.
2) The ACP cache can also receive snoops from the CPU cache hierarchy (to
maintain coherence for the small cache). This does not closely embody the
"halfway coherence" philosophy of ACP.

Due to the above shortcomings in the existing ACP implementation, partially
because of our premature knowledge of the classic memmory model, we use
Ruby instead. In this patch, an ACP write will directly modify data in the
L2 cache (potentially recalling data back from the L1 caches). And an ACP
request won't list the accelerator as a new sharer or owner, therefore no
coherence message will be forwarded to the accelerator.

Change-Id: I6349d1ea46e743020f782e31b97e337ecfd84a0d