Renames all benchmark names in Xenon scripts to use underscores, not hyphens.
A hyphen in the benchmark name results in that name being used to define the
fully qualified statistic name (e.g. system.a-name.tlb.hits) in the Python
world in gem5, but hyphens cannot be part of a Python name.

This is a short term fix for issue #6. A more comprehensive fix would be to
prohibit hyphens entirely or automatically convert them to underscores.

Change-Id: I43f016c2d51125477b22f5639b82310b0617dbc9