systolic-array: Add a systolic array accelerator.
This models a systolic array accelerator, which uses output stationary as its
dataflow. The used data layout is NHWC. Three SRAMs of equal size are used to
store inputs, kernels and outputs, respectively. The following diagram
depicts the dataflow and the mapping from inputs/kernels/outputs to the PE
arrays (a 4x4 array in this example). Assume that the input shape is (1, 32,
32, 8), the kernel shape is (16, 3, 3, 8), with stride size 1 and valid padding.

                             |<----------Weight fold------------>|
                                                         Kernel3
                                                Kernel2     |
                                       Kernel1     |        |
                              Kernel0     |        |        |
                                 |        |        |        |
                                 |        |        |        |
                                 V        V        V        V
  ^        InputWindow0 -->  |--PE00--|--PE01--|--PE02--|--PE03--|
Output    InputWindow1 --->  |--PE10--|--PE11--|--PE12--|--PE13--|
fold     InputWindow2 ---->  |--PE20--|--PE21--|--PE22--|--PE23--|
  V     InputWindow3 ----->  |--PE30--|--PE31--|--PE32--|--PE33--|

The inputs (read from local SRAM) are fed from left column of the array and
pumped towards right, while the top row streams in pixels from the different
kerenls and pump them downwards. Here, every input window is the input pixels
in a convolution (which generates a pixel in an output feature map), therefore
in this case every input window is a region of (1, 3, 3, 8) shape. Every PE
column is responsible for generating an output feature map, of which different
PEs produces adjact output pixels in a single feathure map.

As the output size and the kernel size can be larger than the PE array can
fit, we partition them into 'folds' and iterate over them. In the above
example, the output feature map size is 30x30=900, which will be partitioned
into 900/4=225 output folds. Similarly, the 16 kernels will be partitioned into
16/4=4 weight folds.

The performance model is originally ported from ScaleSim, which generates the
SRAM access trace and estimates the latency for processing the offloaded
convolution task.

TODO: currently we don't have an SRAM model though we do generate SRAM
read/write traces. Every SRAM access takes 1 cycle. Padding is not
implemented yet, for now the model only supports valid padding. Also, since
the performance model is based on an analytical model, the results are not
really computated, though we do use DMA to write the results.

Change-Id: I30554353d8a4db16bb2c126a1e7c1634595d5c74