systolic-array: Make the systolic array simulator cycle-accurate.
This gets rid of the analytical model ported from Scale-Sim and instead develops
a cycle accurate model. We are now clear of the dependencies on Scale-Sim.

In a nutshell, the accelerator contains a PE array located at the center with
data flowing from/to three local scratchpads (input, weight, output) that
encompass it.

More specifically, at the left edge of the PE array, a column of fetch units
feed input data from the input scratchpad, and similarly, a row of fetch units
stream in weight data from the weight scratchpad. In addition, to collect the
finished output data from the PE array and write them back to the output
scratchpad, a column of commit units sit beside the right edge of the PE array.
A FIFO buffer is implemented in each fetch/commit unit in order to avoid
data shortage or congestion. For flow control and bandwidth modelling, three
local buses are used to connect the fetch/commit units to the corresponding
scratchpad. Within the PE array, the input data flow horizontally from left to
right, and the weight data flow vertically from top to bottom.

Thankfully, for generating addresses in the fetch/commit units, the tensor
index interator implemented in SMAUGv2 proves to be a perfect fit (with
necessary adapations), which otherwise would be a really error-prone part and
also a big undertaking.

Also, for modelling the registers in the PEs, which requires data-latching at a
new cycle, the gem5's TimeBuffer class is used (which is used in modelling gem5
CPUs' pipeline inter-stage buffers).

Change-Id: Ib544d095827f0c06c744c0414510c6c53022a9b1