systolic-array: Add an option for accumulating outputs.
This is used when the weight tensor is tiled channelwise, so we need to
accumulate the partial sums across invocations.

The change basically makes the commit unit read the previous partial
sums first (if output accumulation is used), add it to the current output,
and then write the new partial sums back to the scratchpad.

Change-Id: Ibd01ec1a6b688d512497aebfa453765dad5d3707