systolic-array: Pop the commit queue properly.
Previously, the front item of the commit queue was popped as soon as it was
sent, and the total number of outstanding requests was tracked separately. To
model it more properly, we should only pop the front item when it receives the
acknowledge from the scratchpad.

This will also be useful for subsequent commits on adding support for outputs
accumulation.

Change-Id: I599bd84f03adc5779c6806c030fd74d0b71ad93e