systolic-array: Avoid name clashing with SMAUG headers.
The enum and struct used for activations have the same name as those defined in
the SMAUG project. To integrate the invoking of the systolic array into SMAUG,
we have to use different names (note that here we are in the C land, namespace
or enum class can't be used).

Change-Id: Icd3675cf0741a2c44231ce7fe7959025975d56b6