systolic-array: Support different types of datapaths in gem5 config file.
Previously, we used an commondline option (--accel-type) to specify the
accelerator type, which limits the simulation to use only one type of
accelerator. This solves this by adding an accelerator_type option in each
section of the gem5 config file, so we can run Aladdin and the systolic array at
the same time. By default, Aladdin datapath will be used.

Change-Id: Ibd45f1d3fa8d5d42ed3a20987302dc99ce990455