systolic-array: Fix a bug when kernels start from the middle.
When the original weights are tiled, the non-last invocations do not use all the
data stored in the weights scratchpad. To check if a weight unit has finished
reading all the data, we previously used the end() call, which will return
true only when the whole weights tensor has been read, leading to a case where
the invocation has finished all its work whereas the weight units are still
reading data, which should be used by the next invocation. To solve this, use
the number of finished weight folds instead to check whether a weight unit has
finished its work.

Change-Id: Ia7b1a0d2dd73f7d3b3e685710927a1ecd3efd1ef