aladdin: Changes according to the new interrupt-based synchronization.
This adds a special request code for suspending the CPU if the value of the
finish flag is NOT_COMPLETED. A few other changes are made in Aladdin's ioctl
syscall implementation in order to use gem5's BufferArg class for reading values
from simulated memory addresses.

Also the systolic array accelerator now uses the interrupt-based method for
synchronization.

Change-Id: I71a1c0b297ce4f69d12c92c8315554ff5d41e110