aladdin: Fix the sequencer for ACP requests.
The Ruby sequencer locally buffers a request if it aliases with an outstanding
request, where the buffered request will be woken up and treated as a hit once
the outstanding request finishes. This is true for a private cache with
write-allocate policy, because a read/write request will return with a line of
data, and the aliased requests can directly process on that. However, if a
write-through policy is used in the connected controller (as is the case for the
ACP controller), a write request doesn't return with data, thus this
optimization cannot be used.

The fix is instead of buffering aliased requests, simply let the sender retry
later (when the outstanding request finishes, the sender will be notified with
a retry signal).

Change-Id: Ic0313509b061ad7a14c4e5dcf0867281ea5e3e44