stats: Changes due to the new hierarchical stats support.
gem5 upstream now has a new way for structuring stats. The stat system is now
aware of the hierarchical nature of stats: the SimObject hierarchy is explicit
in the stat system. For example, previously we needed to manually specify the
hierarchy of the stat in its name, such as

  acp_latency.name("system." + datapath_name + ".acp_latency")

Now it has a Group class which makes stat naming less fragile and makes it
possible to implement hierarchical formats like XML/HDF5/JSON in a clean way.
For example, we can have a stat group in the HybridDatapath class.

  class HybridDatapath {
    struct HybridDatapathStats : public Stats::Group {
        Stats::Scalar scalar0;
        Stats::Scalar scalar1;

        MyStats(HybridDatapath* dev)
            : Stats::Group(dev),
              scalar0(this, "scalar0", "Description of scalar0"),
              scalar1(this, "scalar1", "Description of scalar1")
        {}
    };
  };

And the user can do m5.stats.dump(datapaths[0]) in Python, which will dump the
stat group that consists of scalar0 and scalar 1 (with the correct stat
hierarchy).

Two virtual functions need to be implemented for every output format:
beginGroup() and endGroup(). These functions keep track of the current stat
hierarchy, making sure when the stat is 'visited' (i.e., a call to visit()), the
correct name of the stat will be printed. For the SQL output format, I basically
copied the way gem5 people implemented these two functions for the text output
format (see the Stats::Text class).

This also updates the Aladdin submodule pointer.

Change-Id: Iac1021045ed13ad650dd4cbc49ef464b4a4c3441