tests: Fix run_travis return code handling properly.
An interrupt should stop the build and return zero without running the
tests. A successful build should go on and run the tests.  Also,
eliminate the retry mechanism. There's no point to that anymore - it
was a relic of debugging the build system from a long time ago.

Finally - just returning a value from main doesn't actually propagate
that as the final return code. We need to sys.exit() too.

Change-Id: I33c561b52e5fa211262bfe3bf88f518103600063