sim-se,cpu-o3: Fix exit syscall for x86 O3 CPU.
When the O3 CPU executes a exit syscall, it doesn't immediately remove the
thread context until a scheduled threadExitEvent is processed. This deferred
exit is added by change e541567, which makes sure that no inflight squash or
trap event will be left incomplete.

One problem with the original change is that it schedules the thread removing
event only when the O3 CPU raises a trap. This works for ISAs such as Arm and
RISC-V, where the execution of the syscall instruction returns a fault (e.g.,
SupervisorCall in Arm), which will be detected by the CPU and causes a trap.
The trap will be handled by the invoke() method of the fault, where it calls
syscall() of the thread (or execution) context which eventually calls the
syscall implementation in Process. But for x86, the process is simplified that
no intentional fault is raised in the execute() method of the 'syscall' or
'int80' instruction, and instead it directly calls xc->syscall(&fault). This
leads to a thread never properly exits.

To fix this, this change schedules a threadExitEvent (if not already scheduled),
when the commit stage sees an exiting thread successfully commits a syscall
instruction.

Change-Id: I453fd0f56487f370c2a95492d1b3f30e6696dfa9